--- Week3/consumer.py
@@ -1,55 +1,0 @@
-import sys
-import os
-
-from confluent_kafka import Consumer, KafkaException, KafkaError
-
-os.environ['CLOUDKARAFKA_BROKERS'] ='tricycle-01.srvs.cloudkafka.com:9094,tricycle-02.srvs.cloudkafka.com:9094,tricycle-03.srvs.cloudkafka.com:9094'
-os.environ['CLOUDKARAFKA_USERNAME'] = "o4zgo7lx"
-os.environ['CLOUDKARAFKA_PASSWORD'] = "x78KarUO8i5ccvEqO0WzEoNR7g7-DIfc"
-os.environ['CLOUDKARAFKA_TOPIC'] = "o4zgo7lx-default"
-
-
-if __name__ == '__main__':
-    topics = os.environ['CLOUDKARAFKA_TOPIC'].split(",")
-
-    # Consumer configuration
-    # See https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md
-    conf = {
-        'bootstrap.servers': os.environ['CLOUDKARAFKA_BROKERS'],
-        'group.id': "%s-consumer" % os.environ['CLOUDKARAFKA_USERNAME'],
-        'session.timeout.ms': 6000,
-        'default.topic.config': {'auto.offset.reset': 'smallest'},
-        'security.protocol': 'SASL_SSL',
-	'sasl.mechanisms': 'SCRAM-SHA-256',
-        'sasl.username': os.environ['CLOUDKARAFKA_USERNAME'],
-        'sasl.password': os.environ['CLOUDKARAFKA_PASSWORD']
-    }
-
-    c = Consumer(**conf)
-    c.subscribe(topics)
-    try:
-        while True:
-            msg = c.poll(timeout=1.0)
-            if msg is None:
-                continue
-            if msg.error():
-                # Error or event
-                if msg.error().code() == KafkaError._PARTITION_EOF:
-                    # End of partition event
-                    sys.stderr.write('%% %s [%d] reached end at offset %d\n' %
-                                     (msg.topic(), msg.partition(), msg.offset()))
-                elif msg.error():
-                    # Error
-                    raise KafkaException(msg.error())
-            else:
-                # Proper message
-                sys.stderr.write('%% %s [%d] at offset %d with key %s:\n' %
-                                 (msg.topic(), msg.partition(), msg.offset(),
-                                  str(msg.key())))
-                print(msg.value())
-
-    except KeyboardInterrupt:
-        sys.stderr.write('%% Aborted by user\n')
-
-    # Close down consumer to commit final offsets.
-    c.close()
