+++ Week3/producer-cloud.py
@@ -1,0 +1,44 @@
+import sys
+import os
+
+from confluent_kafka import Producer
+
+os.environ['CLOUDKARAFKA_BROKERS'] ='192.169.10.106:9092'
+os.environ['CLOUDKARAFKA_USERNAME'] = "kafka"
+os.environ['CLOUDKARAFKA_PASSWORD'] = "EY7BygbW0Vmh"
+os.environ['CLOUDKARAFKA_TOPIC'] = "o4zgo7lx-default"
+
+if __name__ == '__main__':
+    topic = os.environ['CLOUDKARAFKA_TOPIC'].split(",")[0]
+
+    # Consumer configuration
+    # See https://github.com/edenhill/librdkafka/blob/master/CONFIGURATION.md
+    conf = {
+        'bootstrap.servers': os.environ['CLOUDKARAFKA_BROKERS'],
+        'session.timeout.ms': 6000,
+        'default.topic.config': {'auto.offset.reset': 'smallest'},
+        'security.protocol': 'SASL_SSL',
+	'sasl.mechanisms': 'SCRAM-SHA-256',
+        'sasl.username': os.environ['CLOUDKARAFKA_USERNAME'],
+        'sasl.password': os.environ['CLOUDKARAFKA_PASSWORD']
+    }
+
+    p = Producer(**conf)
+
+    def delivery_callback(err, msg):
+        if err:
+            sys.stderr.write('%% Message failed delivery: %s\n' % err)
+        else:
+            sys.stderr.write('%% Message delivered to %s [%d]\n' %
+                             (msg.topic(), msg.partition()))
+
+    for line in sys.stdin:
+        try:
+            p.produce(topic, line.rstrip(), callback=delivery_callback)
+        except BufferError as e:
+            sys.stderr.write('%% Local producer queue is full (%d messages awaiting delivery): try again\n' %
+                             len(p))
+        p.poll(0)
+
+    sys.stderr.write('%% Waiting for %d deliveries\n' % len(p))
+    p.flush()
